refactor early scripts, add more muni/metro brt
